// my-turborepo/apps/admin/src/types/enums.ts
export enum RolUsuario {
  SUPERADMIN = 'SUPERADMIN',
  DUENO = 'DUENO',
  EMPLEADO = 'EMPLEADO',
  CLIENTE = 'CLIENTE',
}

export enum EstadoTurno {
  DISPONIBLE = 'DISPONIBLE',
  RESERVADO = 'RESERVADO',
  SENA_ENVIADA = 'SENA_ENVIADA',
  CONFIRMADO = 'CONFIRMADO',
  CANCELADO = 'CANCELADO',
  EXPIRADO = 'EXPIRADO',
  AUSENTE = 'AUSENTE',
  BLOQUEADO = 'BLOQUEADO',
}

export enum EstadoPago {
  PENDIENTE = 'PENDIENTE',
  ENVIADO = 'ENVIADO',
  APROBADO = 'APROBADO',
  RECHAZADO = 'RECHAZADO',
}

export enum EstadoCancha {
  HABILITADA = 'HABILITADA',
  DESHABILITADA = 'DESHABILITADA',
  EN_MANTENIMIENTO = 'EN_MANTENIMIENTO',
}

export enum MetodoPago {
  TRANSFERENCIA = 'TRANSFERENCIA',
  EFECTIVO = 'EFECTIVO',
  MERCADOPAGO = 'MERCADOPAGO',
  OTRO = 'OTRO',
}